EX00 Finished

+++ CPP00/ex00/megaphone.cpp
@@ -9,7 +9,7 @@ int main(int argc, char const *argv[])
 		std::cout << "* LOUD AND UNBEARABLE FEEDBACK NOISE *";
 	for (int i = 1; i < argc; i++){
 		string = argv[i];
-		for (int j = 0; j < string.length(); j++)
+		for (size_t j = 0; j < string.length(); j++)
 			std::cout << (char) toupper(string[j]);
 	}
 	std::cout << std::endl;
